+++ panther/__init__.py
@@ -1,6 +1,6 @@
 from .main import Panther
 
-__version__ = '1.7.4'
+__version__ = '1.7.5'
 
 
 def version():